backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.10
email-validator==2.2.0
pandas==2.2.3
openpyxl==3.1.5
numpy==2.1.1
Pillow==10.4.0
apscheduler==3.10.4
httpx==0.27.2
python-dotenv==1.0.1
aiofiles==24.1.0
